// Web/Client/src/Auth/auth-helpers/typeKeys.ts
export enum typeKeys {
	LOGIN = "LOGIN",
	LOGIN_SUCCEEDED = "LOGIN_SUCCEEDED",
	LOGIN_FAILED = "LOGIN_FAILED",
	REGISTER = "REGISTER",
	REGISTER_SUCCEEDED = "REGISTER_SUCCEEDED",
	REGISTER_FAILED = "REGISTER_FAILED",
	LOGOUT = "LOGOUT",
	LOGOUT_SUCCEEDED = "LOGOUT_SUCCEEDED",
	LOGOUT_FAILED = "LOGOUT_FAILED",
	AUTHENTICATE_WITH_GOOGLE = "AUTHENTICATE_WITH_GOOGLE",
	AUTHENTICATE_WITH_GOOGLE_SUCCEEDED = "AUTHENTICATE_WITH_GOOGLE_SUCCEEDED",
	AUTHENTICATE_WITH_GOOGLE_FAILED = "AUTHENTICATE_WITH_GOOGLE_FAILED",
	POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE = "POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE",
	POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE_SUCCEEDED = "POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE_SUCCEEDED",
	POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE_FAILED = "POPULATE_USER_STATE_FROM_LOCAL_STORAGE_IF_AVAILABLE_FAILED"
}